refactor: restruct Editable avatar
