+++ src/components/EditableAvatar/AvatarEditModal/IconCollection/IconCollection.module.scss
@@ -1,0 +1,122 @@
+.defaultIconSection {
+  flex: 1 1 auto;
+  padding: 16px 0;
+  .defaultIconContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    width: 100%;
+    .defaultIconHeader {
+      display: flex;
+      align-items: center;
+      padding: 4px 0 16px 0;
+      .backBtn {
+        width: 32px;
+        height: 32px;
+        border-radius: 16px;
+        align-items: center;
+        margin: 0px;
+        max-width: 100%;
+        position: relative;
+        transition: background 0.1s ease-out 0s,
+          box-shadow 0.15s cubic-bezier(0.47, 0.03, 0.49, 1.38) 0s;
+        background: rgba(9, 30, 66, 0.04);
+        cursor: pointer;
+        padding: 0px 2px;
+        outline: none;
+        border: none;
+        color: #42526e;
+        &:hover {
+          transition: background 0 s ease-out 0s,
+            box-shadow 0.15s cubic-bezier(0.47, 0.03, 0.49, 1.38) 0s;
+          background: rgba(9, 30, 66, 0.08);
+        }
+        span {
+          transition: opacity 0.3s ease 0s;
+          opacity: 1;
+          align-self: center;
+          display: flex;
+          flex-grow: 0;
+          flex-shrink: 0;
+          line-height: 0;
+          font-size: 0px;
+          user-select: none;
+          margin: 0px 2px;
+          width: 24px;
+          height: 24px;
+          overflow: hidden;
+          pointer-events: none;
+          max-width: 100%;
+          max-height: 100%;
+          color: rgb(66, 82, 110);
+          vertical-align: bottom;
+          svg {
+            color: rgb(66, 82, 110);
+            font-size: 22px;
+          }
+        }
+      }
+      h4 {
+        padding-left: 8px;
+        color: #091e42;
+        pointer-events: auto;
+
+        font-size: 14px;
+        font-style: normal;
+        font-weight: 500;
+        line-height: 1.43;
+      }
+    }
+    ul {
+      width: 100%;
+      list-style: none;
+      display: grid;
+      grid-template-columns: repeat(4, 1fr);
+      max-height: 294px;
+      overflow-y: auto;
+      padding: 0px;
+      margin: 0px;
+      gap: 8px;
+      li {
+        border-radius: 3px;
+        overflow: hidden;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        width: 100%;
+        border: none;
+        cursor: pointer;
+        margin: 0px;
+        aspect-ratio: 1 / 1;
+        img {
+          object-fit: contain;
+          width: 100%;
+          height: 100%;
+        }
+      }
+    }
+  }
+}
+
+@media (max-width: 320px) {
+  .modalBackdrop {
+    .modalContent {
+      .popupPage {
+        .popupSection {
+          bottom: 30px;
+          .popupWindow {
+            width: 297px;
+            height: 470px;
+            .defaultIconSection {
+              .defaultIconContainer {
+                ul {
+                  width: 318px;
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
