+++ src/components/EditableAvatar/AvatarEditModal/ImageCroper/ImageCroper.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ImageCroper() {
+  return <div>ImageCroper</div>;
+}
+
+export default ImageCroper;
