+++ src/components/EditableAvatar/EditableAvatar.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../variables.scss';
+@import '../../variables.scss';
 
 .icon {
   display: inline-flex;
