+++ src/lib/Modal/Modal.module.scss
@@ -19,6 +19,24 @@
     overflow-y: auto;
     box-shadow: rgb(43 44 57 / 14%) 0px 6px 20px;
     border-bottom: 6px solid #6a2add;
+    position: absolute;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
+    margin: auto;
+    .popupBody {
+      padding: 30px 26px;
+      position: relative;
+      display: flex;
+      flex-direction: column;
+      background-color: #ffffff;
+      color: #091e42;
+      pointer-events: auto;
+      h3 {
+        font-weight: 600;
+        font-size: 20px;
+      }
+    }
   }
 }
